# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photos directly from your PC, and they will be visible to everyone following you. 
    This API allows users to follow / unfollow other users, like and comment on their photos, ban / unban other users and search for their profiles. 
    A user can log in and sign up to their profile with their username, which can also be changed later.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    photoId:
      description: unique photo identifier
      type: integer
      readOnly: true
      example: 123456789
    username:
      description: name of the user profile
      type: string
      pattern: 'ˆ.*?$'
      minLength: 3
      maxLength: 16
      #example: 'Maria'
    PhotoMultipartResponse:
      description: an object that consists of a photo and its related metada
      type: object
      properties:
        photo:
          description: the photo the user wants to upload
          type: string
          format: binary
          minLength: 3
          maxLength: 10000
        caption:
          description: an optional textual caption for the photo
          type: string
          #example: "hello friends"
          minLength: 0
          maxLength: 150
          pattern : "ˆ.*?$"
        location:
          description: an optional location of the photo
          type: string
          #example: "Rome, Italy"
          minLength: 0
          maxLength: 30
          pattern : "ˆ.*?$"
        author:
          $ref: '#/components/schemas/username'
        uploadDateTime:
          description: date and time of the photo upload (ISO 8601 format)
          type: string
          format: date-time
          example: "2022-03-01T12:00:00"
          readOnly: true
          minLength: 4
          maxLength: 30
        photoId:
          $ref: '#/components/schemas/photoId'
        likesCount:
          type: integer
          description: number of likes on the photo
          default: 0
          readOnly: true
        likes:
          $ref: '#/components/schemas/LikeList'
        commentsCount:
          type: integer
          description: number of comments on the photo
          default: 0
          readOnly: true
        comments:
          $ref: '#/components/schemas/CommentList'
...
    Like:
      description: object that describes a like on a photo by a user
      type: object
      properties:
        username:
          $ref: '#/components/schemas/username'
        photoId:
          $ref: '#/components/schemas/photoId'
    CommentResponse:
      description: object that describes a comment on a photo by a user
      type: object
      properties:
        commentId:
          description: unique comment identifier
          type: integer
          readOnly: true
          example: 987654321
        photoId:
          $ref: '#/components/schemas/photoId'
        author:
          $ref: '#/components/schemas/username'
        text:
          description: textual content of the comment    
          type: string
          #example: "Nice photo!"
          minLength: 0
          maxLength: 150
          pattern : "ˆ.*?$"
        date:
          description: date of comment upload
          type: string
          format: date
          example: "2022-08-16"
          minLength: 4
          maxLength: 30
          readOnly: true
...
    Follow:
      description: object that describes a follow on a photo by a user
      type: object
      properties:
        followedUsername:
          $ref: '#/components/schemas/username'
      required:
      - followedUsername
...
    BanRequest:
      description: objects that describes a ban request on a user
      type: object
      required:
        - BannedUsername
      properties:
        BannedUsername:
          $ref: '#/components/schemas/username'
    ConflictResponseBan:
      type: object
      description: object that describes a conflict that emerges when the same ban request already exists
      properties:
        message:
          description: notification to the user that they have already banned that user
          type: string
          example: "User already banned"
          minLength: 1
          maxLength: 100
        existing_ban:
          $ref: '#/components/schemas/BanRequest'
    FollowedUsers:
      description: array of all users that the specified user is following
      type: array
      items:
        {$ref: '#/components/schemas/Follow'}
      minItems: 0
      maxItems: 1000
    PhotoList:
      description: array of photos
      type: array
      items:
        {$ref: '#/components/schemas/PhotoMultipartResponse'}
      minItems: 0
      maxItems: 1000  
    BannedUsers:
      description: array of all users that the specified user has banned
      type: array
      items:
        {$ref: '#/components/schemas/BanRequest'}
      minItems: 0
      maxItems: 1000
    CommentList:
      description: array of all comments on a specific photo
      type: array
      items:
        {$ref: '#/components/schemas/CommentResponse'} 
      minItems: 0
      maxItems: 10000
    LikeList:
      description: array of all likes on a specific photo
      type: array
      items:
        {$ref: '#/components/schemas/Like'} 
      minItems: 0
      maxItems: 10000
    UserProfile:
      description: profile of the user with their photos and data
      type: object
      properties:
        username:
          $ref: '#/components/schemas/username'
        total_photos:
          type: integer
          description: Total number of photos uploaded by the user
          readOnly: true
        followersCount:
          readOnly: true
          type: integer
          description: Number of followers of the user
        followersList:
          $ref: '#/components/schemas/Followers'
        followingCount:
          type: integer
          description: Number of users the user is following
        followingList:
          $ref: '#/components/schemas/FollowedUsers'
        userPhotos:
          $ref: '#/components/schemas/PhotoList'
...
    Followers:
      description: array of all users that follow the specific user
      type: array
      items:
        type: object
        properties:
          followerUsername:
            $ref: '#/components/schemas/username'
      minItems: 0
      maxItems: 10000
...
paths:
...
  /users/{username}/banned/:
    post:
      operationId: banUser
      tags: ["Ban"]
      summary: Ban other users
      description: ban other users so that they can not view your profile, phots, likes and comments, followers and followed users
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          description: username of the user sending a ban request
          required: true
          schema:
            $ref: '#/components/schemas/username'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BanRequest'
              
      responses:
        '201':
          description: User banned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BanRequest'
        '400':
          description: Bad request - check request parameters
        '401':
          description: Unauthorized -user is not authenticated
        '403':
          description: Forbidden - user can not ban themselves, or the username in bearer token does not match the provided path parameter username
        '404':
          description: User not found
        '409':
          description: Conflict - User is already banned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponseBan'
        '500':
          description: Internal server error 
      x-idempotent: true  
      # Custom extension to mark the operation as idempotent- a user sending a ban request for the same user should be the same as banning that user once
    get:
      summary: list banned users
      tags: ["Ban"]
      operationId: getBanned
      description: Retrieve the list of users that have been banned by the user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          description: username of the user requesting the list
          required: true
          schema:
            $ref: '#/components/schemas/username'
      responses:
        '200':
          description: Successful operation - returns a list of banned users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BannedUsers'
        '400':
          description: Bad request - check request parameters
        '401':
          description: Unauthorized -user is not authenticated
        '403':
          description: Forbidden - user can not access other users bannned list, or the username in bearer token does not match the provided path parameter username
        '404':
          description: User not found or has not banned anyone
        '500':
          description: Internal server error
...
  /users/{username}/profile:
    get:
      summary: Get user profile information
      tags: ["User profile"]
      operationId: getProfile
      description: Retrieve information for the user's personal profile page - the user’s photos (in reverse chronological order), how many photos have been uploaded, and the user’s followers and following
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          description: username of the user whose profile is viewed, used also to do a query on /photos collection for those whose author is this username and sort them in descending order by the uploadDateTime paramater
          required: true
          schema:
            $ref: '#/components/schemas/username'
      responses:
        '200':
          description: Successful operation - return user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Bad request - check request parameters
        '401':
          description: Unauthorized - user is not authenticated
        '403':
          description: Forbidden - user is not allowed access because of the ban
        '404':
            description: User not found
        '500':
            description: Internal server error
...